incrowd/requirements.txt
argparse==1.2.1
django-cors-headers==1.0.0
django-filter==0.9.1
django-mailgun==0.2.2
Django==1.7.2
djangorestframework==3.0.2
djoser==0.2.0
MySQL-python==1.2.5
newrelic==2.40.0.34
pusher==0.8
requests==2.5.1
six==1.9.0
supervisor==3.1.3
urllib3==1.10
uWSGI==2.0.9
wsgiref==0.1.2
